started reorganizinng Bot scripts: movemennt script only for moving bot and controller for the rest
