+++ Assets/GameManager.cs
@@ -266,7 +266,7 @@ public class GameManager : NetworkBehaviour
 
         NetworkServer.Spawn(bot.gameObject);
 
-        bot.GetComponent<BotMovement>().SetGraph(graph);
+        bot.GetComponent<BotController>().SetGraph(graph);
         bot.StartBot();
     }
 
@@ -364,11 +364,11 @@ public class GameManager : NetworkBehaviour
             bot.Send();
         }
 
-        // stop all Bot async routines
-        foreach (BotMovement bot in FindObjectsOfType<BotMovement>())
-        {
-            bot.StopEverything();
-        }
+        // // stop all Bot async routines
+        // foreach (BotController bot in FindObjectsOfType<BotController>())
+        // {
+        //     bot.StopEverything();
+        // }
 
         // after countdown go back to lobby
         StartCoroutine(LoadLobbyCountdown());
