+++ Assets/Scripts/Bots/BotController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5e799c12a157b4e86b073898fa879f23
+guid: 93143e2e1ff0b4f579fb38b37363dd18
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
